[SERVICE-578] Review changes
